Adjusted file pathing and run selection for robustness
I removed a '//' preceding a lot of file names and added '.png' calls to figures. This may require rewriting some partner code to correct pathing but allows is in-line with other sources. The run selection is not complete but fixes some issues such as plotting of gsa results if gsa.runSobol= true even if gsa.run =false